requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
anyio==4.3.0
apache-airflow==2.9.1
apache-airflow-providers-common-io==1.3.1
apache-airflow-providers-common-sql==1.13.0
apache-airflow-providers-fab==1.1.0
apache-airflow-providers-ftp==3.9.0
apache-airflow-providers-http==4.11.0
apache-airflow-providers-imap==3.6.0
apache-airflow-providers-smtp==1.7.0
apache-airflow-providers-sqlite==3.8.0
apispec==6.6.1
argcomplete==3.3.0
asgiref==3.8.1
astroid==3.1.0
asttokens==2.4.1
attrs==23.2.0
Babel==2.15.0
blinker==1.8.2
cachelib==0.9.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
comm==0.2.2
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.3
croniter==2.0.5
cryptography==42.0.7
dash==2.17.0
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
dnspython==2.6.1
docutils==0.21.2
email_validator==2.1.1
evdev==1.7.1
EWMHlib==0.2
execnet==2.1.1
executing==2.0.1
flake8==7.0.0
Flask==2.2.5
Flask-AppBuilder==4.4.1
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.6.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2024.3.1
google-re2==1.1.20240501
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.63.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.24.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
lazy-object-proxy==1.10.0
limits==3.12.0
linkify-it-py==2.0.3
lockfile==0.12.2
loguru==0.7.2
Mako==1.3.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdit-py-plugins==0.4.1
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.2.0
MouseInfo==0.1.3
msgspec==0.18.6
multidict==6.0.5
nest-asyncio==1.6.0
notify-py==0.3.42
numpy==1.26.4
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
ordered-set==4.1.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pendulum==3.0.0
pexpect==4.9.0
pip-review==1.3.0
platformdirs==4.2.1
plotly==5.22.0
pluggy==1.5.0
prison==0.2.1
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
PyAutoGUI==0.9.54
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.8.0
pylint==3.1.1
PyMonCtl==0.92
PyMsgBox==1.0.9
pynput==1.7.7
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytest==8.2.0
pytest-xdist==3.6.1
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-multipart==0.0.9
python-nvd3==0.16.0
python-slugify==8.0.4
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.1
PyWinBox==0.7
PyWinCtl==0.4
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.31.0
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rich==13.7.1
rich-argparse==1.4.0
rpds-py==0.18.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
setproctitle==1.3.3
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
SoundCard==0.4.3
SQLAlchemy==1.4.52
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.37.2
tabulate==0.9.0
tenacity==8.3.0
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.14.1
tomlkit==0.12.5
tornado==6.4
traitlets==5.14.3
types-pytz==2024.1.0.20240417
typing_extensions==4.11.0
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.2
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==2.2.3
wheel==0.43.0
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.18.1
